Rename DomainSuggestion component to DomainSuggestionItem
