+++ packages/domain-search/src/components/search-results/placeholder.tsx
@@ -1,4 +1,4 @@
-import { DomainSuggestionsList, DomainSuggestion } from '../../ui';
+import { DomainSuggestionsList, DomainSuggestionItem } from '../../ui';
 
 const PLACEHOLDER_COUNT = 10;
 
@@ -6,7 +6,7 @@ export const SearchResultsPlaceholder = () => {
 	return (
 		<DomainSuggestionsList>
 			{ Array.from( { length: PLACEHOLDER_COUNT } ).map( ( _, index ) => (
-				<DomainSuggestion.Placeholder key={ index } />
+				<DomainSuggestionItem.Placeholder key={ index } />
 			) ) }
 		</DomainSuggestionsList>
 	);
