+++ packages/domain-search/src/ui/domain-suggestion-item/featured.placeholder.tsx
