+++ packages/domain-search/src/ui/domain-suggestion-item/featured.skeleton.tsx
