+++ packages/domain-search/src/ui/domain-suggestion-item/index.skeleton.tsx
