+++ packages/domain-search/src/ui/domain-suggestion-item/style.scss
