+++ packages/domain-search/src/ui/domain-suggestion-item/unavailable.tsx
