+++ packages/domain-search/src/ui/index.ts
@@ -6,7 +6,7 @@ export { DomainsFullCartItems } from './domains-full-cart-items';
 export { DomainsFullCartItem } from './domains-full-cart-items/item';
 export { DomainSuggestionsList } from './domain-suggestions-list';
 export { FeaturedDomainSuggestionsList } from './featured-domain-suggestions-list';
-export { DomainSuggestion } from './domain-suggestion';
+export { DomainSuggestionItem } from './domain-suggestion-item';
 export { DomainSuggestionBadge } from './domain-suggestion-badge';
 export { DomainSuggestionPrice } from './domain-suggestion-price';
 export * from './domain-suggestion-cta';
